rename groupId and pacakges
