+++ src/main/java/com/cloudbees/jenkins/plugins/kubernetes_credentials_provider/convertors/UsernamePasswordCredentialsConvertor.java
@@ -21,13 +21,13 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.cloudbees.jenkins.plugins.k8sCredentialProvider.convertors;
+package com.cloudbees.jenkins.plugins.kubernetes_credentials_provider.convertors;
 
 import io.fabric8.kubernetes.api.model.Secret;
 import hudson.Extension;
-import com.cloudbees.jenkins.plugins.k8sCredentialProvider.CredentialsConvertionException;
-import com.cloudbees.jenkins.plugins.k8sCredentialProvider.SecretToCredentialConverter;
-import com.cloudbees.jenkins.plugins.k8sCredentialProvider.SecretUtils;
+import com.cloudbees.jenkins.plugins.kubernetes_credentials_provider.CredentialsConvertionException;
+import com.cloudbees.jenkins.plugins.kubernetes_credentials_provider.SecretToCredentialConverter;
+import com.cloudbees.jenkins.plugins.kubernetes_credentials_provider.SecretUtils;
 import com.cloudbees.plugins.credentials.CredentialsScope;
 import com.cloudbees.plugins.credentials.impl.UsernamePasswordCredentialsImpl;
