+++ src/main/java/com/cloudbees/jenkins/plugins/kubernetes_credentials_provider/package-info.java
@@ -26,4 +26,4 @@
  */
 @edu.umd.cs.findbugs.annotations.ReturnValuesAreNonnullByDefault
 @edu.umd.cs.findbugs.annotations.DefaultAnnotation(edu.umd.cs.findbugs.annotations.NonNull.class)
-package com.cloudbees.jenkins.plugins.k8sCredentialProvider;
+package com.cloudbees.jenkins.plugins.kubernetes_credentials_provider;
