+++ src/test/java/com/cloudbees/jenkins/plugins/kubernetes_credentials_provider/convertors/CertificateCredentialsConvertorTest.java
@@ -21,13 +21,13 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.cloudbees.jenkins.plugins.k8sCredentialProvider.convertors;
+package com.cloudbees.jenkins.plugins.kubernetes_credentials_provider.convertors;
 
 import java.io.InputStream;
-import java.nio.charset.StandardCharsets;
+import java.security.KeyStore;
+import java.security.cert.X509Certificate;
 import io.fabric8.kubernetes.api.model.Secret;
 import io.fabric8.kubernetes.client.utils.Serialization;
-import org.jenkinsci.plugins.plaincredentials.impl.FileCredentialsImpl;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -39,21 +39,25 @@ import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
 import hudson.util.HistoricalSecrets;
 import jenkins.security.ConfidentialStore;
-import com.cloudbees.jenkins.plugins.k8sCredentialProvider.CredentialsConvertionException;
+import com.cloudbees.jenkins.plugins.kubernetes_credentials_provider.CredentialsConvertionException;
+import com.cloudbees.jenkins.plugins.kubernetes_credentials_provider.convertors.CertificateCredentialsConvertor;
 import com.cloudbees.plugins.credentials.CredentialsScope;
+import com.cloudbees.plugins.credentials.impl.CertificateCredentialsImpl;
 import static org.hamcrest.CoreMatchers.containsString;
 import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.CoreMatchers.allOf;
 import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.fail;
 
 /**
- * Tests FileCredentialsConvertor
+ * Tests CertificateCredentialsConvertor
  */
 @RunWith(PowerMockRunner.class)
 @PrepareForTest({ConfidentialStore.class, HistoricalSecrets.class})
 @PowerMockIgnore({"javax.crypto.*" }) // https://github.com/powermock/powermock/issues/294
-public class FileCredentialsConvertorTest {
+public class CertificateCredentialsConvertorTest {
+
 
     @Before
     public void mockConfidentialStore() {
@@ -67,102 +71,132 @@ public class FileCredentialsConvertorTest {
 
     @Test
     public void canConvert() throws Exception {
-        FileCredentialsConvertor convertor = new FileCredentialsConvertor();
-        assertThat("correct registration of valid type", convertor.canConvert("secretFile"), is(true));
+        CertificateCredentialsConvertor convertor = new CertificateCredentialsConvertor();
+        assertThat("correct registration of valid type", convertor.canConvert("certificate"), is(true));
         assertThat("incorrect type is rejected", convertor.canConvert("something"), is(false));
     }
 
     @Test
     public void canConvertAValidSecret() throws Exception {
-        FileCredentialsConvertor convertor = new FileCredentialsConvertor();
+        CertificateCredentialsConvertor convertor = new CertificateCredentialsConvertor();
 
         try (InputStream is = get("valid.yaml")) {
             Secret secret = Serialization.unmarshal(is, Secret.class);
             assertThat("The Secret was loaded correctly from disk", notNullValue());
-            FileCredentialsImpl credential = convertor.convert(secret);
+            CertificateCredentialsImpl credential = convertor.convert(secret);
             assertThat(credential, notNullValue());
-            assertThat("credential id is mapped correctly", credential.getId(), is("another-test-file"));
-            assertThat("credential description is mapped correctly", credential.getDescription(), is("secret file credential from Kubernetes"));
+            assertThat("credential id is mapped correctly", credential.getId(), is("a-test-certificate"));
+            assertThat("credential description is mapped correctly", credential.getDescription(), is("certificate credential from Kubernetes"));
             assertThat("credential scope is mapped correctly", credential.getScope(), is(CredentialsScope.GLOBAL));
-            assertThat("credential username is mapped correctly", credential.getFileName(), is("mySecret.txt"));
-            assertThat("credential password is mapped correctly", credential.getSecretBytes().getPlainData(), is("Hello World!".getBytes(StandardCharsets.UTF_8)));
+            assertThat("credential password is mapped correctly", credential.getPassword().getPlainText(), is("testPassword"));
+            KeyStore ks = credential.getKeyStore();
+            // credential.getKeyStore never returns null so we need to check the Keystore contains our certificate
+            assertThat("credential certificate mapped correctly ", ks.containsAlias("myKey"), is(true));
+            // TODO what can we check here to see this is valid
+            X509Certificate cert  = (X509Certificate) ks.getCertificate("myKey");
+            assertThat("Correct cert", cert.getSubjectDN().getName(), is("CN=A Test, OU=Dev, O=CloudBees, L=Around The World, ST=Cool, C=earth"));
         }
     }
 
     @Test
-    public void canConvertAValidMappedSecret() throws Exception {
-        FileCredentialsConvertor convertor = new FileCredentialsConvertor();
+    public void canConvertAValidSecretWithMapping() throws Exception {
+        CertificateCredentialsConvertor convertor = new CertificateCredentialsConvertor();
 
         try (InputStream is = get("validMapped.yaml")) {
             Secret secret = Serialization.unmarshal(is, Secret.class);
             assertThat("The Secret was loaded correctly from disk", notNullValue());
-            FileCredentialsImpl credential = convertor.convert(secret);
+            CertificateCredentialsImpl credential = convertor.convert(secret);
             assertThat(credential, notNullValue());
-            assertThat("credential id is mapped correctly", credential.getId(), is("another-test-file"));
-            assertThat("credential description is mapped correctly", credential.getDescription(), is("secret file credential from Kubernetes"));
+            assertThat("credential id is mapped correctly", credential.getId(), is("a-test-certificate"));
+            assertThat("credential description is mapped correctly", credential.getDescription(), is("certificate credential from Kubernetes"));
             assertThat("credential scope is mapped correctly", credential.getScope(), is(CredentialsScope.GLOBAL));
-            assertThat("credential username is mapped correctly", credential.getFileName(), is("mySecret.txt"));
-            assertThat("credential password is mapped correctly", credential.getSecretBytes().getPlainData(), is("Hello World!".getBytes(StandardCharsets.UTF_8)));
+            assertThat("credential password is mapped correctly", credential.getPassword().getPlainText(), is("testPassword"));
+            KeyStore ks = credential.getKeyStore();
+            // credential.getKeyStore never returns null so we need to check the Keystore contains our certificate
+            assertThat("credential certificate mapped correctly ", ks.containsAlias("myKey"), is(true));
+            // TODO what can we check here to see this is valid
+            X509Certificate cert  = (X509Certificate) ks.getCertificate("myKey");
+            assertThat("Correct cert", cert.getSubjectDN().getName(), is("CN=A Test, OU=Dev, O=CloudBees, L=Around The World, ST=Cool, C=earth"));
         }
     }
- 
+
     @Test
-    public void failsToConvertWhenFilenameMissing() throws Exception {
-        FileCredentialsConvertor convertor = new FileCredentialsConvertor();
+    public void failsToConvertWhenCertificateMissing() throws Exception {
+        CertificateCredentialsConvertor convertor = new CertificateCredentialsConvertor();
         
-        try (InputStream is = get("missingFilename.yaml")) {
+        try (InputStream is = get("missingCertificate.yaml")) {
             Secret secret = Serialization.unmarshal(is, Secret.class);
             convertor.convert(secret);
             fail("Exception should have been thrown");
         } catch (CredentialsConvertionException cex) {
-            assertThat(cex.getMessage(), containsString("missing the filename"));
+            assertThat(cex.getMessage(), containsString("missing the certificate"));
         }
     }
 
     
     @Test
-    public void failsToConvertWhenDataMissing() throws Exception {
-        FileCredentialsConvertor convertor = new FileCredentialsConvertor();
+    public void failsToConvertWhenPasswordMissing() throws Exception {
+        CertificateCredentialsConvertor convertor = new CertificateCredentialsConvertor();
 
-        try (InputStream is = get("missingData.yaml")) {
+        try (InputStream is = get("missingPassword.yaml")) {
             Secret secret = Serialization.unmarshal(is, Secret.class);
             convertor.convert(secret);
             fail("Exception should have been thrown");
         } catch (CredentialsConvertionException cex) {
-            assertThat(cex.getMessage(), containsString("missing the data"));
+            assertThat(cex.getMessage(), containsString("missing the password"));
         }
     }
 
+    /**
+     * Tests when the certificate is not a valid certificate
+     */
     @Test
-    public void failsToConvertWhenFilenameCorrupt() throws Exception {
-        FileCredentialsConvertor convertor = new FileCredentialsConvertor();
+    public void failsToConvertWhenCertificateCorruptPKCS12() throws Exception {
+        CertificateCredentialsConvertor convertor = new CertificateCredentialsConvertor();
         
-        try (InputStream is = get("corruptFilename.yaml")) {
+        try (InputStream is = get("corruptCertificatePKCS12.yaml")) {
             Secret secret = Serialization.unmarshal(is, Secret.class);
             convertor.convert(secret);
             fail("Exception should have been thrown");
         } catch (CredentialsConvertionException cex) {
-            assertThat(cex.getMessage(), containsString("invalid filename"));
+            assertThat(cex.getMessage(), allOf(containsString("invalid certificate"), containsString("PKCS#12")));
         }
     }
 
-    
+
+    /**
+     * Tests when the certificate is not valid base64
+     */
     @Test
-    public void failsToConvertWhenDataCorrupt() throws Exception {
-        FileCredentialsConvertor convertor = new FileCredentialsConvertor();
-
-        try (InputStream is = get("corruptData.yaml")) {
+    public void failsToConvertWhenCertificateCorruptBase64() throws Exception {
+        CertificateCredentialsConvertor convertor = new CertificateCredentialsConvertor();
+        
+        try (InputStream is = get("corruptBase64Certificate.yaml")) {
             Secret secret = Serialization.unmarshal(is, Secret.class);
             convertor.convert(secret);
             fail("Exception should have been thrown");
         } catch (CredentialsConvertionException cex) {
-            assertThat(cex.getMessage(), containsString("invalid data"));
+            assertThat(cex.getMessage(), allOf(containsString("invalid certificate"), containsString("base64")));
         }
     }
+
+    @Test
+    public void failsToConvertWhenPasswordCorrupt() throws Exception {
+        CertificateCredentialsConvertor convertor = new CertificateCredentialsConvertor();
+
+        try (InputStream is = get("corruptPassword.yaml")) {
+            Secret secret = Serialization.unmarshal(is, Secret.class);
+            convertor.convert(secret);
+            fail("Exception should have been thrown");
+        } catch (CredentialsConvertionException cex) {
+            assertThat(cex.getMessage(), containsString("invalid password"));
+        }
+    }
+
 
     @Test
     public void failsToConvertWhenDataEmpty() throws Exception {
-        FileCredentialsConvertor convertor = new FileCredentialsConvertor();
+        CertificateCredentialsConvertor convertor = new CertificateCredentialsConvertor();
         
         try (InputStream is = get("void.yaml")) {
             Secret secret = Serialization.unmarshal(is, Secret.class);
@@ -174,7 +208,7 @@ public class FileCredentialsConvertorTest {
     }
 
     private static final InputStream get(String resource) {
-        InputStream is = FileCredentialsConvertorTest.class.getResourceAsStream("FileCredentialsConvertorTest/" + resource);
+        InputStream is = CertificateCredentialsConvertorTest.class.getResourceAsStream("CertificateCredentialsConvertorTest/" + resource);
         if (is == null) {
             fail("failed to load resource " + resource);
         }
