+++ src/test/java/com/cloudbees/jenkins/plugins/kubernetes_credentials_provider/convertors/StringCredentialCredentialsConvertorTest.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.cloudbees.jenkins.plugins.k8sCredentialProvider.convertors;
+package com.cloudbees.jenkins.plugins.kubernetes_credentials_provider.convertors;
 
 import java.io.InputStream;
 import io.fabric8.kubernetes.api.model.Secret;
@@ -29,7 +29,8 @@ import io.fabric8.kubernetes.client.utils.Serialization;
 import static org.junit.Assert.*;
 import org.jenkinsci.plugins.plaincredentials.impl.StringCredentialsImpl;
 import org.junit.Test;
-import com.cloudbees.jenkins.plugins.k8sCredentialProvider.CredentialsConvertionException;
+import com.cloudbees.jenkins.plugins.kubernetes_credentials_provider.CredentialsConvertionException;
+import com.cloudbees.jenkins.plugins.kubernetes_credentials_provider.convertors.StringCredentialConvertor;
 import com.cloudbees.plugins.credentials.CredentialsScope;
 import static org.hamcrest.CoreMatchers.*;
