+++ src/test/resources/com/cloudbees/jenkins/plugins/kubernetes_credentials_provider/convertors/FileCredentialsConvertorTest/corruptFilename.yaml
